Make nn package independent on data package
